Move pytest_addoption to src
